Add metric filtering information to README
Prometheus has the ability to filter which metrics it sends to its
remote write target. Users can set this up to lower space usage and
improve performance in the remote storage solution. An additional use
case is only sending metrics of a certain type, e.g., numerical, to a
given long-term storage solution.